Move manifest filegroup inside of cargo_buildscript_rule instead of separate target
Summary:
This filegroup does not need to be its own separate `native.filegroup` with its own dedicated target name. It can just be constructed inside the implementation of the buildscript rule using `ctx.actions.symlinked_dir`.

The advantage of the new way is it simplifies D73625771 and D73625775 by avoiding needing to invent a unique target name for this filegroup in the case of local path dependencies.

Reviewed By: diliop

Differential Revision: D73625773

fbshipit-source-id: 30f30bd6da815919a40c06aa35218ab1147a5702