Include Kotlin version in incremental_metadata.proto.json
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This diff
In D72721609 we added a mechanism to force a full recompilation on a config change. This diff adds kotlin version to the list.

Differential Revision: D73592340

fbshipit-source-id: 9d874240efb0dcc0345de5283d276c1aed5f1a32