Preserve compilation options for debugging
Summary:
# Context
The buck2 rules use `+deterministic` to strip  absolute paths from the beam info, etc. This has the side effect that all options are removed from the `compile` metadata of the beam file.

Because there are tools that recompile modules dynamically and preserve the options (e.g. meck), we want to preserve some options, like `+line_coverage`

# Problem
At the moment, we are not preserving debugging options like `debug_info` and `beam_debug_info` (new in OTP/28).

# This diff
Add support for both

Reviewed By: michalmuskala

Differential Revision: D73768158

fbshipit-source-id: 96c7b66f34e88de7c4d7eebeeec9bb93452424cf