Use adb_server_port if provided
Summary: RL users use OD with ssh tunnelling to connect to local devices attached to their laptop or leased via RE. In their setup they may run `adb` on a specific port and provide it to `buck install`. We should provide that port to every `adb` call via `-P port` to work which is what this diff does.

Reviewed By: christolliday

Differential Revision: D73923358

fbshipit-source-id: 00787b6559653379732325421814444ab147a015