Implementation of duplicated class check
Summary:
Sometimes people run into the problem of the "Class X is defined multiple times" error,
but it only happens for the opt builds, which are usually slow to build.

Since we already have the information of all the classes that are going to be packaged
into the apk, we can use this information to check for any duplicates and provide the
signal much earlier.

Reviewed By: navidqar

Differential Revision: D73590794

fbshipit-source-id: edc7c04409ed1d52c831cf8240318b42e70229c7