Support override to statically link deps of shared-only nodes
Summary:
Currently, if omnibus encounters a shared-only linkable in the graph,
it'll apply this shared-only linkage transitively to all it's deps, as it's
possible that the shared-only node is prebuilt and  contains `DT_NEEDED`
tags to it's deps.

This diff provides a way for nodes to override this behavior in the case
this heuristic isn't desired.

Reviewed By: technicianted

Differential Revision: D74024273

fbshipit-source-id: e016802ea0a70b11456f486dda773e15a8dae624