erlang: directly build the include directory
Summary:
This switches us to using `ctx.actions.copied_dir` rather than individually copying each include file.
This has the advantage that we can get the path for the final directory straght-away and we don't need to first build a fake directory with an anchor.

Furthermore, we can just symlink the include directory into the final "linked"  path, rather than building a directory with individual symlinked files.

In the `dep` calculation the `file` we're trying to access has to be the same basename as `artifact.basename`, so we can just use that instead

Reviewed By: TD5, TheGeorge

Differential Revision: D74086040

fbshipit-source-id: 025d2db2a0416fab12d5c9bcf7e7c963e50448e3