erlang: improve provider typing
Summary:
Use precise types for providers (except application, there because of the multi-toolchain builds types end up more complex and improvements are left for later).

We also don't need the separation between the `ErlangToolchainInfo` provider and `Toolchain` record - we can just re-use the provider as a record.
I didn't rename all the use-sites and instead added a type alias to avoid large refactoring of the entire erlang prelude codebase.

Reviewed By: TD5

Differential Revision: D74312513

fbshipit-source-id: 7abf1cc824bc8b5a20d7b9c6f18f9d61ece111d8