Avoid creating large frozen defs in darwin dist_lto logic
Summary: Apparently the [dynamic_output API](https://buck2.build/docs/api/build/AnalysisActions/#analysisactionsdynamic_output) we use here generates bytecode for unique call to the function, injecting captured values into the produced bitcode. This means if we make N `dynamic_output` calls for N object files in our link, we allocate N copies of the bytecode for the called function. We are going to have to do this no matter what, but we can make this function as small as possible, by delegating to a function that does the heavy lifting though a small lambda.

Reviewed By: rmaz

Differential Revision: D74506642

fbshipit-source-id: 6209c5c64f2291feebb3e71ba758c6843d568d9c