erlang: Remove input mapping
Summary:
This seems to be not necessary for correctness of builds themselves. We only need the mapping for include files, and we can store that separately for each include application, removing the massive union `dict`.

This was used to detect duplicates in includes, however the detection on `dpes` files happens earlier in the code and already coveres all the conflict cases. This improves error messages coming from that and adds tests that we correctly issue errors on conclicts.

This takes down retained memory down to 3.2GB (reduction by 31%) and total allocations to 8.9GB (reduction by 20%).

Reviewed By: TheGeorge

Differential Revision: D74642907

fbshipit-source-id: 0983516d3a856fce11fbb2b9b663541189e412a0