erlang: store includes and private includes the same way
Summary:
This changes the storage of private includes to be exactly the same as regular includes. This avoids building the masisve dict of private includes and sames some memory since we can store the include directory separately instad of with each file.

This has a potential performance downside - private incldues are now searched linearly in applications instead of direct lookup. However, for normal applications, we'll have just one private include directory, and for tests where we peek includes from dependencies, we generally don't have many includes, so this shouldn't be a big issue in practice. I couldn't measure performance impact of this change.

This slightly regresses memory retained by ~100MB due to extra storage on `ErlangAppInfo`.

Reviewed By: TD5

Differential Revision: D74657860

fbshipit-source-id: ca8671dcba55ca89c23ef97c922da9df42b09bec