erlang: less dependencies when building .app files
Summary:
We don't need all the sources for modules when we build the `.app` file - just the template. This should slightly decrease memory use and reduce the data transfer to RE - we don't need to send all the source files when building the `.app` file

Having the output path as separate argument makes this easier to code.

Reviewed By: TD5

Differential Revision: D74824626

fbshipit-source-id: 2d9b61ee7c3719dec21182df14074329fb652e05