Pass Clang to dist_lto_opt action wrapper in non positional argument
Summary:
D74704259 makes it awkward to pass the compiler to this script as a positional argument. In fixing that, I discovered that this positional argument was not used anyways.

This diff introduces an named parameter for the compiler, and correctly uses that Clang instead of the Clang included with the linker toolchain component.

Reviewed By: rmaz

Differential Revision: D74831527

fbshipit-source-id: 4028d11b70a901087378ba8d8c2d2f31afb41fa3