Ensure explicit targets have the correct mode hashes
Summary:
The buck out mode hashes are dynamic now and we can't use one sample target to find the hashes for all the targets now.
This is reported by the avatar team in https://fb.workplace.com/groups/1496591530773407/posts/2217146652051221/?comment_id=2221295408303012
What I'm doing in this stack of diffs is to make sure that at least the explicitly listed targets have the correct mode hashes.
This diff calls the bxl scirpt in last diff on the explicit targets from command line arguments

Reviewed By: autozimu

Differential Revision: D75199081

fbshipit-source-id: 66abee472001d1ce8cca0fd1341696957094b8e2