bxl script to get mode hashes for explicit targets
Summary:
The buck out mode hashes are dynamic now and we can't use one sample target to find the hashes for all the targets now.
This is reported by the avatar team in https://fb.workplace.com/groups/1496591530773407/posts/2217146652051221/?comment_id=2221295408303012
What I'm doing in this stack of diffs is to make sure that at least the explicitly listed targets have the correct mode hashes.
To do that we need to query the output path of them all and `buck build` would be very slow. This diff adds a new bxl script to get the mode hashes of targets in bxl.

Reviewed By: autozimu

Differential Revision: D75199083

fbshipit-source-id: dd67e3be4ddf6fc6e45138789c7cf7cccffad2de