Support language level 9+
Summary:
Now that we have a JDK system image, wire it up for `target >= 9` Android usages.

I still use the `CustomJdkInfo.bootclasspath` field with `target >= 9` as a convenient way to propagate android.jar to the `-classpath`. In the distant future when support for java8 release is dropped, this field should be renamed.

The motivation for doing this is primarily to unblock Mockito 5.X in our codebase which requires Java 11 bytecode: https://github.com/mockito/mockito/releases/tag/v5.0.0

Reviewed By: navidqar

Differential Revision: D75542820

fbshipit-source-id: 17870e7993ed7c46a807be644ef7e7117f999bd6