Add script to create JDK system image from Android core-for-system-modules.jar
Summary:
See the top comment in the script. Basically I am porting this feature from Gradle:
https://cs.android.com/android-studio/platform/tools/base/+/mirror-goog-studio-main:build-system/gradle-core/src/main/java/com/android/build/gradle/internal/dependency/JdkImageTransformDelegate.kt

With `javac` we currently use `-bootclasspath=android.jar` to replace "system" classes with the Android jar. This flag only works with `target=8`, and in fact all Java-based rules in fbandroid default to `target = 8` still.

To support target=9+, we need to transform `core-for-system-modules.jar` from the Android SDK to a custom JDK image and pass this to the `--system` flag.

This will enable using `source/target=11` for Android libraries with Java sources.

Reviewed By: navidqar

Differential Revision: D75468761

fbshipit-source-id: 959da06be917d06099ff4ccc5f659c44c9a14ddc