Remove kotlinc_run_via_build_tools_api config
Summary:
In the previous diff, I removed the JarBackedReflectedKotlinc implementation as we have adopted BuiltToolsApi as the default entry point for compiler interaction.

This diff continues the cleanup process by removing the kotlinc_run_via_build_tools_api configuration, which is now redundant and has no effect following the removal of JarBackedReflectedKotlinc.

⚠️ Land only if D75772131 reaches stable with no issue ⚠️

Reviewed By: rybalkinsd

Differential Revision: D75850236

fbshipit-source-id: 61d06d870023b3824323f10ac951c32c3d357328