Rollback "Make compiler_flags be of arg type"
Summary: The change to make compiler_Flags be of arg type broke the ide.bxl script that we use to drive ghcide. My later attempt at fixing it by converting the args to string was not sucessful, and is not the right way to address this. So I am rolling back the change for now, open to future changes around this area as long as they also address the ide.bxl script

Differential Revision: D75959871

fbshipit-source-id: daa63873e312563bfc7df18fcc431f0997c81b55