Extract dirs from resource manifest
Summary: When defining resources users are able to pass in directories. For each library this checks each resource and adds it's contents to the manifest. Previously this was done in make_par and was an expensive part of par building.  This change guarantees that every item in the manifest is a file.

Reviewed By: christycylee

Differential Revision: D75911735

fbshipit-source-id: f176fdd915c8e6898cd441eba3811b68353e1399