Move thin-link into separate file
Summary:
Until now, all the logic for a distributed link is implemented in a single function with number of closures capturing a tremendous amount of state. The function is getting too large, and in particular the rampant variable capture makes it hard to reason about where state is mutated.

To make things a little clearer in preparation for double codegen, move thin-link to its own function, in its own file. We will need to pass a large number of parameters, but at least we are clear about what state is required and might be mutated.

Reviewed By: rmaz

Differential Revision: D75806417

fbshipit-source-id: 9c98596201941918d5b24bc2c3685b21fa79af83