Remove extra Darwin distributed thin-link linker argfiles from [linker.command] subtarget
Summary: Previously information was added to the JSON file [linker.command] subtarget to include arguments used for a couple commands run during a distributed link. This was meant to use be used in tests to compare against local thin-lto as distributed thin-lto for Darwin support was introduced. Now that we have flag parity, the argsfiles provide limited utility now, and complicate the rules.

Reviewed By: rmaz

Differential Revision: D75806407

fbshipit-source-id: c10cc6737b6ae650548f0a3b90d540743a17d8b9