Place all flags in the argsfile for thin-link as well
Summary: We have somewhat arbitrarily distributed some flags for thin link in a response file, while leaving others on the command line. We currently place common_link_flags in the command directly rather than in the argsfile. In a future change, we want to include more flags in common_link_flags. To avoid making the command too verbose when that happens, include as much as possible in the argsfile.

Reviewed By: rmaz

Differential Revision: D75806425

fbshipit-source-id: c7601ef1f858356be5b45ff75ba92bd2a3ea35e1