Use experimental content-based paths in _generate_build_config_dot_java
Summary:
This is a very simple action that we can use to begin to test content-based paths. It takes in a single written file of defaults and produces a `BuildConfig.java` file that is compiled.

All the pieces that we need for content-based hashing to work for this action should be in place.

Reviewed By: siaojiecai

Differential Revision: D76043657

fbshipit-source-id: e12ce91ef5d64aa4a262d61c160c55559b51d37f