Refactor ignore_artifacts in _mk_hmap
Summary:
Right now, we add `header_args` as `hidden` to `generate_hmap`, but we also ignore all artifacts on `header_args`, so we're not actually adding anything to `generate_hmap`. So, remove `hidden = header_args,`.

We can then get rid of `ignore_artifacts = True`, since `write` will always ignore artifacts anyway (unless they are content-based, in which case we need those artifacts to not be ignored so that we can calculate the correct path to write).

Reviewed By: nhatminhle

Differential Revision: D76129120

fbshipit-source-id: c8e704114ed971e71ea856eca202d4bd886507b3