Use better zstd/avoid deflate.
Summary:
For language packs in IG and threads we were using zstd compression but then they were defalted in the ipa.
Generally, it is not efficient to compress things twice and we can see that the deflate compression was not efficient (<2%, see P1825371501)

I made a post about it (https://fb.workplace.com/groups/igiosqa/permalink/3908801659334897/ ) but could not find any owners so not sure who could review and stamp this.

Reviewed By: lacbs

Differential Revision: D75661456

fbshipit-source-id: f45dc07620ceb88882d66d714c38439511d01bff