accept project name as keys in `versioned_*` params
Summary:
This diff is to add temporary compatibility for a different format of `versioned_*` fields in targets like `cpp_library_external`. Here is an example of what these parameters look like:
https://www.internalfb.com/code/fbsource/[2ce87c0fe3a93982101e1246f7912538b654f701]/fbcode/third-party-buck/platform010/build/cudnn/9.2.0/TARGETS?lines=21-34

With this diff, the above example can now be specified as:
```
versioned_header_dirs = [
    (
        {"cuda": "latest"},
        ["76a58c6/include"],
    ),
    (
        {"cuda": "12.4"},
        ["a6270e7/include"],
    ),
    (
        {"cuda": "12"},
        ["e32163f/include"],
    ),
],
```

The code that parses these params in `_extract_versions` only extracts the project name, so we can update all usages of these params to have project name as the inner dictionary keys, and it seems like all usages of these params exist within the `third-party-buck` targets generated by `buckify_tp2`.

This is both to make the `buckify_tp2` rewrite slightly easier, plus it's one of the final cleanups for the `__project__` target removed in D74862454.

The next diff will update `buckify_tp2` to output `versioned_*` params in the new format, then finally remove compatibility for parsing `:__project__` targets from `_extract_versions`.

NOTE: we do plan to clean this up eventually so that `buckify_tp2` outputs the proper buck2 selects, instead of having additional logic in the prelude layer to build the selects. For the sake of finishing the `buckify_tp2` rust rewrite sooner, we'll do it later.

Reviewed By: igorsugak

Differential Revision: D76327000

fbshipit-source-id: db84b0b7d573c6d7ce38c0976f25361a7188d526